+++ data/validation_split_ids/gnormplus.txt
@@ -1,0 +1,63 @@
+8867793
+17889706
+2208279
+9473669
+8182091
+8167223
+17160896
+10484778
+10753822
+7506095
+8663294
+12052141
+9662397
+19370769
+8751386
+1317062
+9215638
+17627763
+10639512
+18805828
+8663110
+10915768
+8632143
+11875050
+1698610
+7806211
+19450126
+7479798
+9215685
+8204608
+9161407
+2786251
+14755728
+10707983
+7556196
+9480844
+7957066
+16737910
+12749065
+3858849
+9766636
+1329093
+7683036
+9027582
+7957065
+17223398
+20596603
+9346895
+12117771
+8555498
+9931490
+10441506
+10072425
+8243272
+1693656
+7556175
+7925474
+1651173
+2674141
+7624799
+18991055
+12215838
+18406868
