+++ data/validation_split_ids/nlm_gene.txt
@@ -1,0 +1,68 @@
+24004819
+20821041
+27460078
+27748834
+26847345
+24402617
+18507837
+28882895
+27473374
+21951951
+28393246
+25277705
+28395336
+25412315
+28260077
+19201842
+23775080
+12805933
+18310445
+27167772
+28350299
+25818835
+28846666
+23651211
+27779699
+28416580
+28259974
+26655640
+26782292
+26545401
+30462722
+28403862
+27943262
+19825989
+26890356
+26963473
+27470968
+26476672
+28115395
+28259936
+25305591
+26640231
+27509880
+26450454
+28084316
+28228553
+23195221
+24694877
+28260110
+21810259
+26027517
+28186140
+25622904
+17612403
+23460868
+22022612
+28348168
+18673534
+21152263
+28534510
+24743235
+30442153
+28359316
+20926645
+25946186
+23220172
+23069675
+28428259
